Add files via upload
a missile system warning simulator inspired by Hawaii false warning due to its bad UI design. the control of flow of information. will be a simple absurd website reimagining the possible variations of bad ui design choices (and obtrusive ads). or just a 're-production' of the missle interface based on conflicting information from the us government regarding details of the failure. need to implement more js for this last part